feature/1-Implement-clean-architecture - initial setup for clean architecture
