+++ Flashcards.Api/FlashcardRequestParams.cs
@@ -1,11 +1,9 @@
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Flashcards.Api;
 
-public class Flashcard : BaseEntity
+public class FlashcardRequestParams
 {
-    [ForeignKey("DeckId")]
     public int DeckId { get; set; }
 
     [Required] public string Question { get; set; } = null!;
