+++ Flashcards.Api/Flashcards.Api.csproj
@@ -11,8 +11,8 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="8.0.11" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.1">
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.2" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.2">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
@@ -24,6 +24,11 @@
   <ItemGroup>
     <ProjectReference Include="..\Flashcards.Application\Flashcards.Application.csproj" />
     <ProjectReference Include="..\Flashcards.Contracts\Flashcards.Contracts.csproj" />
+    <ProjectReference Include="..\Flashcards.Infrastructure\Flashcards.Infrastructure.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Compile Remove="BaseEntity.cs" />
   </ItemGroup>
 
 </Project>
