--- Flashcards.Application/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Flashcards.Application;
-
-public class Class1
-{
-
-}
