--- Flashcards.Contracts/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Flashcards.Contracts;
-
-public class Class1
-{
-
-}
