--- Flashcards.Domain/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Flashcards.Domain;
-
-public class Class1
-{
-
-}
