+++ Flashcards.Domain/Flashcard.cs
@@ -1,0 +1,14 @@
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace Flashcards.Domain;
+
+public class Flashcard : BaseEntity
+{
+    [ForeignKey("DeckId")]
+    public int DeckId { get; set; }
+
+    [Required] public string Question { get; set; } = null!;
+
+    [Required] public string Answer { get; set; } = null!;
+}
