--- Flashcards.Infrastructure/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Flashcards.Infrastructure;
-
-public class Class1
-{
-
-}
