+++ Flashcards.Infrastructure/Flashcards.Infrastructure.csproj
@@ -2,6 +2,17 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Flashcards.Application\Flashcards.Application.csproj" />
+    <ProjectReference Include="..\Flashcards.Domain\Flashcards.Domain.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.2" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="9.0.2">
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+      <PrivateAssets>all</PrivateAssets>
+    </PackageReference>
+    <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="9.0.3" />
   </ItemGroup>
 
   <PropertyGroup>
